Move ResponseOnlyTxInfo to common file
