+++ packages/xrpl/src/models/methods/accountTx.ts
@@ -1,7 +1,7 @@
-import { LedgerIndex } from '../common'
+import { LedgerIndex, ResponseOnlyTxInfo } from '../common'
 import { Transaction, TransactionMetadata } from '../transactions'
 
-import { ResponseOnlyTxInfo, BaseRequest, BaseResponse } from './baseMethod'
+import { BaseRequest, BaseResponse } from './baseMethod'
 
 /**
  * The account_tx method retrieves a list of transactions that involved the
