+++ packages/xrpl/src/models/methods/subscribe.ts
@@ -1,13 +1,15 @@
-import type { Amount, Currency, Path, StreamType } from '../common'
+import type {
+  Amount,
+  Currency,
+  Path,
+  StreamType,
+  ResponseOnlyTxInfo,
+} from '../common'
 import { Offer } from '../ledger'
 import { OfferCreate, Transaction } from '../transactions'
 import { TransactionMetadata } from '../transactions/metadata'
 
-import type {
-  ResponseOnlyTxInfo,
-  BaseRequest,
-  BaseResponse,
-} from './baseMethod'
+import type { BaseRequest, BaseResponse } from './baseMethod'
 
 interface Book {
   /**
